Move directory handling into link provider
